+++ CnCTDRAMapEditor/Properties/Settings.Designer.cs
@@ -683,6 +683,24 @@ namespace MobiusEditor.Properties {
         [global::System.Configuration.ApplicationScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool TdHelisSpawnOnGround {
+            get {
+                return ((bool)(this["TdHelisSpawnOnGround"]));
+            }
+        }
+        
+        [global::System.Configuration.ApplicationScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool RaHelisSpawnOnGround {
+            get {
+                return ((bool)(this["RaHelisSpawnOnGround"]));
+            }
+        }
+        
+        [global::System.Configuration.ApplicationScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
         public bool FixConcretePavement {
             get {
                 return ((bool)(this["FixConcretePavement"]));