+++ CnCTDRAMapEditor/TiberianDawn/GameInfoTibDawn.cs
@@ -93,6 +93,7 @@ namespace MobiusEditor.TiberianDawn
         public override int MaxTeams => Constants.MaxTeams;
         public override int HitPointsGreenMinimum => 127;
         public override int HitPointsYellowMinimum => 63;
+        public override bool LandedHelis => Globals.LandedHelisTd;
         public override Size ViewportSizeSmall => new Size(240, 192);
         public override Size ViewportSidebarSmall => new Size(80, 192);
         public override Point ViewportOffsetSmall => new Point(0, 0);